ms-analyse-trafic/pyproject.toml
[project]
name = "ms-analyse-trafic"
version = "0.1.0"
description = ""
authors = [
    {name = "fehiZ77",email = "rajonhsonfehizoro@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.136.0,<0.137.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "kafka-python (>=2.2.15,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-cov (>=6.2.1,<7.0.0)"
]

[tool.poetry]
package-mode = false

[tool.pytest.ini_options]
pythonpath = ["src"]